+++ abdm/models/__init__.py
@@ -1,4 +1,5 @@
 from .abha_number import *  # noqa
 from .consent import *  # noqa
 from .health_facility import *  # noqa
+from .inbound_callback import *  # noqa
 from .transaction import *  # noqa
