+++ abdm/service/v3/callback_handlers/__init__.py
@@ -1,0 +1,3 @@
+class CallbackProcessingError(Exception):
+    """Raised when an inbound callback cannot be processed (e.g. referenced
+    entity missing). Marks the envelope FAILED after retries are exhausted."""
