+++ abdm/signals/register_care_contexts.py
@@ -1,12 +1,8 @@
-import logging
-
 from django.db import transaction
-from django.db.models import Q
 from django.db.models.signals import post_save, pre_save
 from django.dispatch import receiver
 
 from abdm.service.helper import (
-    ABDMAPIException,
     create_diagnostic_report_care_context,
     create_encounter_care_context,
     create_file_upload_care_context,
@@ -15,18 +11,15 @@ from abdm.service.helper import (
     create_questionnaire_response_care_context,
     hf_id_from_encounter,
 )
-from abdm.service.v3.gateway import GatewayService
+from abdm.tasks.link_care_context import enqueue_link_care_context
 from care.emr.models.diagnostic_report import DiagnosticReport
 from care.emr.models.encounter import Encounter
 from care.emr.models.file_upload import FileUpload
 from care.emr.models.invoice import Invoice
 from care.emr.models.medication_request import MedicationRequest
-from care.emr.models.observation import Observation
 from care.emr.models.questionnaire import QuestionnaireResponse
 from care.emr.resources.file_upload.spec import FileTypeChoices
 from care.emr.resources.invoice.spec import InvoiceStatusOptions
-
-logger = logging.getLogger(__name__)
 
 
 @receiver(post_save, sender=MedicationRequest)
@@ -49,24 +42,14 @@ def create_care_context_on_medication_request_creation(
     ):
         return
 
-    try:
-        transaction.on_commit(
-            lambda: GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [create_medication_request_care_context(instance)],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_medication_request_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
         )
-    except ABDMAPIException as e:
-        warning = f"Failed to link care context for medication request {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-        logger.warning(warning)
-
-    except Exception as e:
-        warning = f"Failed to link care context for medication request {instance.external_id} with patient {patient.external_id}, {e!s}"
-        logger.exception(warning)
+    )
 
 
 @receiver(post_save, sender=Encounter)
@@ -84,24 +67,14 @@ def create_care_context_on_encounter_creation(
     ):
         return
 
-    try:
-        transaction.on_commit(
-            lambda: GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [create_encounter_care_context(instance)],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_encounter_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
         )
-    except ABDMAPIException as e:
-        warning = f"Failed to link care context for encounter {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-        logger.warning(warning)
-
-    except Exception as e:
-        warning = f"Failed to link care context for encounter {instance.external_id} with patient {patient.external_id}, {e!s}"
-        logger.exception(warning)
+    )
 
 
 @receiver(pre_save, sender=FileUpload)
@@ -128,24 +101,14 @@ def create_care_context_on_file_upload_creation(sender, instance: FileUpload, **
     if not patient or not hf_id or getattr(patient, "abha_number", None) is None:
         return
 
-    try:
-        transaction.on_commit(
-            lambda: GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [create_file_upload_care_context(instance)],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_file_upload_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
         )
-    except ABDMAPIException as e:
-        warning = f"Failed to link care context for file upload {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-        logger.warning(warning)
-
-    except Exception as e:
-        warning = f"Failed to link care context for file upload {instance.external_id} with patient {patient.external_id}, {e!s}"
-        logger.exception(warning)
+    )
 
 
 @receiver(post_save, sender=QuestionnaireResponse)
@@ -163,39 +126,17 @@ def create_care_context_on_questionnaire_response_creation(
     ):
         return
 
-    # Observations are bulk_created after this signal fires, so the check is deferred to on_commit when the related rows are visible in the DB.
-    def link_if_has_coded_observations():
-        has_coded_observation = (
-            Observation.objects.filter(questionnaire_response=instance)
-            .filter(
-                Q(main_code__isnull=False) & ~Q(main_code={})
-                | Q(alternate_coding__isnull=False) & ~Q(alternate_coding=[])
-            )
-            .exists()
+    # Observations are bulk_created after this signal fires; the task checks for
+    # coded observations once the transaction is committed and the rows are visible.
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_questionnaire_response_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
+            questionnaire_response_external_id=str(instance.external_id),
         )
-
-        if not has_coded_observation:
-            return
-
-        try:
-            GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [
-                        create_questionnaire_response_care_context(instance)
-                    ],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
-        except ABDMAPIException as e:
-            warning = f"Failed to link care context for questionnaire response {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-            logger.warning(warning)
-        except Exception as e:
-            warning = f"Failed to link care context for questionnaire response {instance.external_id} with patient {patient.external_id}, {e!s}"
-            logger.exception(warning)
-
-    transaction.on_commit(link_if_has_coded_observations)
+    )
 
 
 @receiver(post_save, sender=DiagnosticReport)
@@ -213,24 +154,14 @@ def create_care_context_on_diagnostic_report_creation(
     ):
         return
 
-    try:
-        transaction.on_commit(
-            lambda: GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [create_diagnostic_report_care_context(instance)],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_diagnostic_report_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
         )
-    except ABDMAPIException as e:
-        warning = f"Failed to link care context for diagnostic report {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-        logger.warning(warning)
-
-    except Exception as e:
-        warning = f"Failed to link care context for diagnostic report {instance.external_id} with patient {patient.external_id}, {e!s}"
-        logger.exception(warning)
+    )
 
 
 @receiver(pre_save, sender=Invoice)
@@ -253,21 +184,11 @@ def create_care_context_on_invoice_issue(sender, instance: Invoice, **kwargs):
     if not patient or not hf_id or getattr(patient, "abha_number", None) is None:
         return
 
-    try:
-        transaction.on_commit(
-            lambda: GatewayService.link__carecontext(
-                {
-                    "patient": patient,
-                    "care_contexts": [create_invoice_care_context(instance)],
-                    "user": instance.created_by,
-                    "hf_id": hf_id,
-                }
-            )
+    transaction.on_commit(
+        lambda: enqueue_link_care_context(
+            patient_external_id=str(patient.external_id),
+            care_context=create_invoice_care_context(instance),
+            hf_id=hf_id,
+            user_id=instance.created_by_id,
         )
-    except ABDMAPIException as e:
-        warning = f"Failed to link care context for invoice {instance.external_id} with patient {patient.external_id}, {e.detail!s}"
-        logger.warning(warning)
-
-    except Exception as e:
-        warning = f"Failed to link care context for invoice {instance.external_id} with patient {patient.external_id}, {e!s}"
-        logger.exception(warning)
+    )
