+++ abdm/tasks/__init__.py
@@ -3,7 +3,9 @@ import random
 from celery import current_app
 from celery.schedules import crontab
 
+from abdm.tasks.link_care_context import link_care_context
 from abdm.tasks.patient_share import patient_share_on_share
+from abdm.tasks.process_inbound_callback import process_inbound_callback
 from abdm.tasks.retry_failed_care_contexts import retry_failed_care_contexts
